Extracts login related functions from the API module.
Login helpers are not related to API calls
