+++ ts/screens/authentication/IdpLoginScreen.tsx
@@ -4,7 +4,6 @@ import { WebView } from "react-native";
 import { NavigationScreenProp, NavigationState } from "react-navigation";
 import { connect } from "react-redux";
 import { ReduxProps } from "../../actions/types";
-import { extractLoginResult } from "../../api";
 import AppHeader from "../../components/ui/AppHeader";
 import * as config from "../../config";
 import I18n from "../../i18n";
@@ -14,6 +13,7 @@ import {
   isUnauthenticatedWithoutIdpSessionState,
   SessionState
 } from "../../store/reducers/session";
+import { extractLoginResult } from "../../utils/login";
 type ReduxMappedProps = {
   session: SessionState;
 };
